Collected incidents: 2017-2022 (Failure w/ Temp=1)

+++ failures/commands/classifyAnalyzable.py
@@ -49,6 +49,7 @@ class ClassifyAnalyzableCommand:
         parser.add_argument(
             "--temp",
             type=float,
+            default=0,
             help="Sets the temperature for ChatGPT",
         )
 
@@ -60,22 +61,28 @@ class ClassifyAnalyzableCommand:
             args (argparse.Namespace): The parsed command-line arguments.
             parser (argparse.ArgumentParser): The argument parser for the configuration
         """
-        
+        logging.info("\nClassifying articles on whether they contain information for software failure analysis.")
+
         # Gets list of article to classify
         queryset = (
             Article.objects.filter(describes_failure=True, id__in=args.articles) if args.articles else
             Article.objects.filter(describes_failure=True) if args.all else
             Article.objects.filter(describes_failure=True, analyzable_failure=None)
         )
-        logging.info("\nClassifying articles on whether they contain enough information for software failure analysis.")
+        
         
         # Initializes ChatGPT Classifier
         classifierChatGPT = ClassifierChatGPT()
 
         # Handles inputs and temperature
-        temperature = args.temp if args.temp else 1
-        inputs = {"model": "gpt-3.5-turbo", "temperature": temperature}
-        logging.info("Classifying articles using temperature %s.", "{:.2f}".format(temperature))
+        if 0 <= args.temp <= 1:
+            temperature = args.temp
+        else:
+            logging.info("\nTemperature out of range [0,1]. Please check input.")
+            exit()
+        
+        inputs = {"model": "gpt-3.5-turbo", "temperature": temperature} #gpt-3.5-turbo , gpt-4-1106-preview
+        logging.info("\nUsing " + inputs["model"] + " with a temperature of " + str(temperature) + ".")
         
         analyzable_positive_classifications_ChatGPT = 0
 
@@ -87,10 +94,12 @@ class ClassifyAnalyzableCommand:
                 analyzable_positive_classifications_ChatGPT += 1
                 logging.info("ChatGPT Classifier: Classification met as eligible for failure analysis for article: " + str(article))
 
+
         logging.info("ChatGPT successfully classified %d articles as eligible for failure analysis.", analyzable_positive_classifications_ChatGPT)
 
-        logging.info("Cleaning up database")
         self.process_incident()
+
+        
 
     def process_incident(self):
         """
@@ -99,42 +108,47 @@ class ClassifyAnalyzableCommand:
 
         Args:
         """
-        logging.info("Resetting incidents/incident relationships for analyzable articles")
         
         # Get list of articles that are not analyzable and do not have an incident tied to them
         articles = Article.objects.filter(analyzable_failure=False, incident__isnull=False)
 
-        # Init vector DB
-        chroma_client = chromadb.HttpClient(host="172.17.0.1", port="8001") #TODO: host.docker.internal
-        embedding_function = OpenAIEmbeddings()
-        vectorDB = Chroma(client=chroma_client, collection_name="articlesVDB", embedding_function=embedding_function)
+        if articles:
 
-        for article in articles:
+            logging.info("Cleaning up database")
 
-            # Get associated incident
-            incident = article.incident
+            logging.info("Resetting incidents/incident relationships for analyzable articles")
 
-            # Check if there is only one article associated with the incident
-            if incident.articles.count() == 1 and incident.articles.first() == article:
-                # Delete from vectorDB
-                logging.info("Deleting incident " + str(incident.id) + " from Django and VectorDB.")
-                chunks_for_incident = vectorDB.get(where={"incidentID": incident.id})['ids']
-                if chunks_for_incident:
-                    vectorDB._collection.delete(ids=chunks_for_incident)
+            # Init vector DB
+            chroma_client = chromadb.HttpClient(host="172.17.0.1", port="8001") #TODO: host.docker.internal
+            embedding_function = OpenAIEmbeddings()
+            vectorDB = Chroma(client=chroma_client, collection_name="articlesVDB", embedding_function=embedding_function)
 
-                # Delete incident from Django
-                incident.delete()
+            for article in articles:
 
-            # Delete article from vectorDB
-            logging.info("Deleting article " + str(article.id) + " from VectorDB")
-            chunks_for_sampleArticle = vectorDB.get(where={"articleID": article.id})['ids']
-            if chunks_for_sampleArticle:
-                vectorDB._collection.delete(ids=chunks_for_sampleArticle)
+                # Get associated incident
+                incident = article.incident
 
-            # Set article incident to none, set article stored to false
-            article.incident = None
-            article.article_stored = False
-            article.save()
+                # Check if there is only one article associated with the incident
+                if incident.articles.count() == 1 and incident.articles.first() == article:
+                    # Delete from vectorDB
+                    logging.info("Deleting incident " + str(incident.id) + " from Django and VectorDB.")
+                    chunks_for_incident = vectorDB.get(where={"incidentID": incident.id})['ids']
+                    if chunks_for_incident:
+                        vectorDB._collection.delete(ids=chunks_for_incident)
+
+                    # Delete incident from Django
+                    incident.delete()
+
+                # Delete article from vectorDB
+                logging.info("Deleting article " + str(article.id) + " from incident " + str(article.incident.id) + " from VectorDB")
+                chunks_for_sampleArticle = vectorDB.get(where={"articleID": article.id})['ids']
+                if chunks_for_sampleArticle:
+                    vectorDB._collection.delete(ids=chunks_for_sampleArticle)
+
+                # Set article incident to none, set article stored to false
+                article.incident = None
+                article.article_stored = False
+                article.save()
